server/app/requirements.txt
Django==5.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
transformers==4.49.0
sentence-transformers==2.7.0
huggingface-hub==0.35.2
django-cors-headers==4.3.0